Added routes to the backend to improve structure, and began working on the chat functionality
