+++ server/controllers/chatController.js
@@ -1,0 +1,9 @@
+const db = require("../db/dbConfig");
+
+const chatController = {
+    create: async (req, res) => {
+        
+    }
+}
+
+  module.exports = chatController;
