+++ server/routes/authRoutes.js
@@ -1,0 +1,10 @@
+const express = require("express");
+const router = express.Router();
+const authController = require("../controllers/authController");
+const limiter = require("../utils/rateLimitConfig.js");
+
+router.post("/login", limiter, authController.login);
+router.post("/register", authController.register);
+
+
+module.exports = router;
