+++ server/utils/validateEmailConfig.js
@@ -1,0 +1,6 @@
+function validateEmail(email) {
+    const emailRegex = /^[^\s@]+@[^\s@]+\.[^\s@]+$/;
+    return emailRegex.test(email);
+}
+
+module.exports = validateEmail;
